requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
filelock==3.13.1
filetype==1.2.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.58
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.42
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.37.1
networkx==3.3
numba==0.61.2
numpy==2.1.2
# openai==1.77.0
openai-whisper==20240930
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.30.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.45.0
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
# torch==2.7.0+cpu
# torchaudio==2.7.0+cpu
# torchvision==0.22.0+cpu
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yt-dlp==2025.4.30
zstandard==0.23.0
